java

+++ src/day25/BankAccount.java
@@ -1,14 +1,14 @@
 package day25;
 
 public class BankAccount {
-	private double balance; // 帳戶餘額
+	private int balance; // 帳戶餘額
 	
-	public BankAccount(double intialBalance) {
+	public BankAccount(int intialBalance) {
 		this.balance = intialBalance;
 	}
 	
 	// 提款方法
-	public void withdraw(double amount) {
+	public void withdraw(int amount) {
 		String threadName = Thread.currentThread().getName();
 		System.out.printf("%s 要提款 $%d\n", threadName, amount);
 		
@@ -19,6 +19,7 @@ public class BankAccount {
 		} else {
 			System.out.printf("%s 提款 $%d 失敗\n", threadName, amount);
 		}
+		
 	}
 	
 }
